feat: add redis config
Using Redis for caching trending post is more efficient than using cache-manager.
Because Redis has rank functionality, it can easily handle the ranking of posts based on their popularity.
And it is more scalable for larger datasets, which is efficient for our use case, 100~200 amount of post data could be large.

- add redis config
- add redis docker compose service
- refactor config structure
- remove cache-manager and use redis directly